[FIX] board user check

+++ ssauction_back/src/main/java/com/ssafy/ssauction/service/board/BoardServiceImpl.java
@@ -3,16 +3,23 @@ package com.ssafy.ssauction.service.board;
 
 import com.ssafy.ssauction.domain.board.Board;
 import com.ssafy.ssauction.domain.board.BoardRepository;
+import com.ssafy.ssauction.domain.board.BoardType;
+import com.ssafy.ssauction.domain.comments.Comments;
 import com.ssafy.ssauction.domain.comments.CommentsRepository;
+import com.ssafy.ssauction.domain.users.Users;
+import com.ssafy.ssauction.domain.users.UsersRepository;
 import com.ssafy.ssauction.web.dto.board.BoardDtoReq;
 import com.ssafy.ssauction.web.dto.board.BoardDtoRes;
+import com.ssafy.ssauction.web.dto.comments.CommentsDtoRes;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.List;
 
 @Service
 public class BoardServiceImpl implements BoardService {
@@ -22,20 +29,23 @@ public class BoardServiceImpl implements BoardService {
 
     private final CommentsRepository commentsRepository;
 
+    private final UsersRepository usersRepository;
+
     @Autowired
-    public BoardServiceImpl(BoardRepository boardRepository, CommentsRepository commentsRepository) {
+    public BoardServiceImpl(BoardRepository boardRepository, CommentsRepository commentsRepository, UsersRepository usersRepository) {
         this.boardRepository = boardRepository;
         this.commentsRepository = commentsRepository;
+        this.usersRepository = usersRepository;
     }
 
     @Override
-    public Page<Board> boardList(PageRequest pageRequest, String searchType, String search) {
+    public Page<Board> boardList(PageRequest pageRequest, BoardType boardType, String searchType, String search) {
         if(searchType == null)
-            return boardRepository.findAll(pageRequest);
+            return boardRepository.findByBoardType(pageRequest, boardType);
         else if("title".equals(searchType))
-            return boardRepository.findByBoardTitleContaining(pageRequest, search);
+            return boardRepository.findByBoardTitleContainingAndBoardType(pageRequest, search, boardType);
         else
-            return boardRepository.findByUserNickname(pageRequest, search);
+            return boardRepository.findByUserNicknameAndBoardType(pageRequest, search, boardType);
     }
 
     @Override
@@ -52,7 +62,27 @@ public class BoardServiceImpl implements BoardService {
             boardDtoRes.setBoardReply(board.getBoardReply());
             boardDtoRes.setUserNo(board.getUserNo());
             boardDtoRes.setUserNickname(board.getUserNickname());
-            boardDtoRes.setComments(commentsRepository.findByBoardNo(boardNo));
+
+            List<CommentsDtoRes> commentsRes = new ArrayList<>();
+            List<Comments> comments = commentsRepository.findByBoardNo(boardNo);
+
+            for (Comments comment : comments) {
+                CommentsDtoRes commentRes = new CommentsDtoRes();
+
+                commentRes.setCommentNo(comment.getCommentNo());
+                commentRes.setCommentContent(comment.getCommentContent());
+                commentRes.setCommentRegDate(comment.getCommentRegDate());
+                commentRes.setBoardNo(comment.getBoardNo());
+                commentRes.setUserNo(comment.getUserNo());
+
+                Users user = usersRepository.findById(commentRes.getUserNo()).orElse(null);
+                if(user != null) {
+                    commentRes.setUserNickname(user.getUserNickname());
+                    commentsRes.add(commentRes);
+                }
+            }
+
+            boardDtoRes.setComments(commentsRes);
         }
         return boardDtoRes;
     }
@@ -81,6 +111,8 @@ public class BoardServiceImpl implements BoardService {
         if(board == null)
             return false;
         else {
+            if(board.getUserNo() != boardDtoReq.getUserNo())
+                return false;
             board.setBoardTitle(boardDtoReq.getBoardTitle() + "(수정됨)");
             board.setBoardContent(boardDtoReq.getBoardContent());
             boardRepository.save(board);
@@ -89,12 +121,14 @@ public class BoardServiceImpl implements BoardService {
     }
 
     @Override
-    public boolean boardDelete(long boardNo) {
+    public boolean boardDelete(long boardNo, long userNo) {
         Board board = boardRepository.findById(boardNo).orElse(null);
 
         if(board == null)
             return false;
         else {
+            if(board.getUserNo() != userNo)
+                return false;
             boardRepository.delete(board);
             return true;
         }
